--- gencfg-asm/src/bin/ease.rs
@@ -1,25 +1,0 @@
-use gencfg_asm::{asm::Assembler, parse_asm};
-use std::fs;
-use std::path::PathBuf;
-use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-#[structopt(name = "assembler")]
-struct Opt {
-    #[structopt(parse(from_os_str))]
-    input: PathBuf,
-}
-
-fn main() {
-    let opt = Opt::from_args();
-    let asm = fs::read_to_string(opt.input).unwrap();
-    let parsed = match parse_asm(&asm) {
-        Ok(p) => p,
-        Err(e) => panic!("Parse error: {:?}", e),
-    };
-    let mut assembler = Assembler::new();
-    assembler.push_all(parsed).unwrap();
-    let output = assembler.take();
-    assembler.finish().unwrap();
-    println!("{}", hex::encode(output));
-}
