--- gencfg-cli/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "gencfg-cli"
-version = "0.1.0"
-authors = ["Sam Wilson <sam.wilson@mesh.xyz>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-hex = "0.4.2"
-structopt = "0.3.21"
-
-[dev-dependencies]
-hex-literal = "0.3.1"
